Fix or suppress Sonar-reported issues

+++ src/main/java/org/xbill/DNS/utils/hexdump.java
@@ -41,7 +41,7 @@ public class hexdump {
           sb.append('\t');
         }
       }
-      int value = (int) b[i + offset] & 0xFF;
+      int value = b[i + offset] & 0xFF;
       sb.append(hex[value >> 4]);
       sb.append(hex[value & 0xF]);
       sb.append(' ');
